BUGFIX: Can't send new version 1.0 for review
The external_edit_summary is required, and is set when new measures are
created (as "First published").  However, because the field is not
rendered at all in the template (not even as hidden) this default value
is overwritten with None when the page is first saved.  And then the
page can't be sent for review because external_edit_summary is empty,
and it can't be set by the user because the box doesn't exist in the
page for 1.0 measures.

This fixes things by always including the form fields for edit notes in
the page, but hidden for 1.0 measures.

The associated data migration here will backfill the required field for
any measures currently created that have had their external_edit_summary
overwritten already.