BUGFIX: Only attempt to delete charts and tables if they exist
Many of our charts and tables do not yet have an associated entry in the
dimension_chart or dimension_table tables, as these are currently only
created when a new classification is saved.

This means that if users try to delete a chart or table from a dimension
it will error if the table_id or chart_id is null.

This fixes things to check that the chart_id or table_id exists before
attempting to delete it.