Consistently quote the names of things in flash confirmation messages
Sometimes the names of deleted things have quotes round them in our
"deleted" confirmation messages and sometimes they don't.  This makes
it so all names have typographical quotes around them, and we use
measure titles rather than URIs as the naes of things, as this is more
likely to make sense to users.