Ensure all POSTed forms include a CSRF token
We now enforce CSRF protection on all POST requests, so this commit
ensures that all forms that currently POST have a CSRF token included.

This also makes it so that the CSRF token is the first thing after the
opening <form> tag, to make it easy to see that the form is CSRF
protected.