Add some explanatory error messages to error pages
As our publish/unpublish/send-to-review pages could be requested
directly with a GET there's a chance that someone might try to do that
when the page isn't in the right state for the requested status change.

Up until now they would have just seen the message "Bad request", with
no hint as to what was wrong with it.

This adds some error messaging that explains the specific problem with
the current request.