Drop temporary table at end of migration
When alembic applies database migrations, it does so with a single
wrapping transaction (by default), which means that temporary tables
persist across individual migrations. Therefore, to keep allow multiple
migrations to use a temporary table either as part of a set of
migrations going in at once, or by itself, we need to drop the temporary
table at the end of each migration that uses it (and create it in any
subsequent migrations that need it).