Add ordering to page service query
We currently query the database and use the returned records to power a
number of our views, but often expect those records to be ordered
consistently. Without an `order by` phrase, the returned order is
arbitrary. Given that we can afford the DB overhead of ordering all our
queries by default, let's do this until it becomes a problem (which
shouldn't be for a while, if ever).

 ## Ticket
https://trello.com/c/c7uMXeps