Use `back_populates` over `backref` on models
When declaring relationships between models in SQLAlchemy, one can
either use 'backref' or 'back_populates'. `backref` allows the
relationship to be fully-realised by only declaring it on one of the
models, whereas `back_populates` requires entries on both models to set
up bi-directional population/access of the related model.

We think it's more explicit, and therefore quicker to reason about, if
you can look at a single model and see all of its relationships, so we
are switching over to use `back_populates` by default.

The main exception to this is the `parent` relationship on the
`MeasureVersion` table (which will eventually disappear), as this is a
relationship to the same table, so using `back_populates` would need us
to declare the relationship twice.

 ## Ticket
https://trello.com/c/GM3GAlys