Improve notification for stalled build
When a build stalls we want to be notified quickly and codify how to
acknowledge the issue to prevent repeated (hourly) notifications. We
also want to have the issue logged in our central error reporting
system, Sentry, where that's configured (currently just production).

This patch adds code that will log the exception with Sentry, if
available, or just allow the exception to bubble up and kill the
process, which should be caught by our logging alerts if we don't have
Sentry.

 ## Ticket
https://trello.com/c/Juc8hRqG