Clean DB before every test session
After the restructure, a few of our tests are complaining about data
already existing in the database which conflicts with data the test
expects/has added itself. This means we have some tests that are
leaking data (i.e. not cleaning up after themselves correctly). The
catch-all solution here is to clean the database before _every_ test,
which we do here with the `autouse=True` parameter for `pytest.fixture`

We also stop refreshing materialized views inside the `db_session`
fixture and push this requirement back into the test which depends on
the materialized views. It is misleading to refresh materialized views
within this fixture as, at the point of refreshing them, there is no
data in the database. The materialized views need to be refreshed
_after_ inserting all of the data required by the test, therefore it
needs to be done inside the test.