Don't manually prescribe ID for Data Source stub
Serial columns in Postgres don't automatically increment when you assign
an ID explicitly, so if you add a record with `id=1` manually and then
add a new record without specifying an id, postgres might also retrieve
`1` from the column's sequence generator.

We could theoretically have the same problem on the `MeasureVersion`
table, as we manually specify ids for those records in our fixtures. I
haven't fixed that here because it doesn't seem to be a problem right
now, and we will have to re-write all those fixtures soon as part of the
sitemap refactor anyway.