Use `pre-commit` for hook management
We currently use a pre-commit hook for enforcing code formatting via the
`black` Python code formatter. We've had problems rolling this out in a
way that works for all developers as we all write code and commit
changes with different software.

The pre-commit framework is well-established and supports multiple
platforms/processes out of the box. I've tested it in
CLI/PyCharm/Sourcetree and it's worked seamlessly each time, so this
looks like a good candidate for drop-in replacement.

https://pre-commit.com/

All that you should need to do to setup the new hooks is to install pip
requirements, remove the old hook, and install the new ones with the
`pre-commit` CLI.

```
pip install -r requirements-test.txt
rm ./.git/hooks/pre-commit
pre-commit install --install-hooks
```

The four hooks that I've installed are:
1) Strip whitespace from the end of lines in edited files
2) Ensure there's a blank line at the end of edited files (to keep Kev
happy)
3) Run `flake8` over edited files for basic static code analysis+linting.
4) Run `black` over edited files for standardised code formatting.