Remove some jquery (#907)
Replace some jQuery with plain DOM methods. Long-term aim is to be able to remove jQuery as a dependency.

This commit mostly replaces these calls:

```jquery
$('#data-panel').hide()
```

with

```jquery
document.getElementById('data-panel').classList.add('hidden')
```

which is slightly longer, but will at least throw an error if the element can't be found.