Add `get_latest_version_of_all_measures` method
We often need to get the latest versions of our measure pages, and with
our new data model this is a trivial query, so the new page service can
have this as part of its interface. We include options to
include/exclude drafts (i.e. unpublished versions), as this is a filter
that is frequently used throughout the site (e.g. for dashboards or
listing all measures in order to share them with a department user).