Implement `create_measure` for new page service
Retire the `create_page` method on the old `page_service` and implement
a new version, using the new tables, on `new_page_service`.

This also adds a `strip_whitespace` argument to the RDUStringField,
which replaces some hacky functionality contained in the `create_page`
and `update_page` methods of the old `page_service`, and applies this
functionality to all of the fields in the MeasurePageForm. It also
refactors how existing filters are added to RDU Form Fields after
reading about a potential memory leak with the current implementation
(https://stackoverflow.com/a/48588271).