Stop using `page_service.get_latest_version`
This wraps up the changes in static_site views, which no longer use the
old `page_service` and are now fully migrated to the `new_page_service`.

It fixes a bug in the old `page_service.create_copy` that would make a
new measure for every copy, including version increments.

We also change the logic of `MeasureVersion.get_versions` to look at its
measure id, rather than the `MeasureVersion.guid`