Move `create_copy` to `new_page_service`
Retire the `create_copy` method on the old `page_service` and create an
equivalent on the `new_page_service`, `create_new_measure_version`. This
can equally create new versions, or copies, of measures. This method
still contains a reasonable amount of logic due to the difficulty in
selecting which fields are eligible to be copied across for a new draft.

In order to avoid throwing around strings like "copy", "minor", and
"major", we create a `NewVersionType` enum that contains entries for
each of these, and pass that into the `create_new_measure_version`
method.

We move away from using `db.session.expunge`/`db.session.make_transient`
on the `measure_version` object in favour of a `dictalchemy`-based
`CopyableModel` mixin for the `MeasureVersion` class. This provides
helper methods that can convert a class instance into a Python dict, and
create a new instance from scratch by consuming said dict.

The `copy_uploads` method on the `upload_service` has also had a
facelift, and is now `copy_uploads_between_measure_version` - taking two
measure version: from (source) and to (dest).