Update NewVersionForm and remove unique macro
With the new `NewVersionType` enum, the new version page was rendering
the same (major) version for both options. I opted to take this time to
clean up the form, as it's carrying some tech debt around a unique form
macro and not using the RDU form fields. We also get proper styling on
the radio buttons now that we're using the consistent fields, and are
able to stop using magic strings for minor/major version changes.