Remove dead code and refactor form
We no longer need to build the england/wales/scotland/northern ireland
variables and pass those into the MeasurePageForm as they're unused. So
let's remove those. At the same time, this refactors the same form to
not need models passed in as kwargs; instead the forms grab the data
directly from the database. This is simpler without really being any
worse - beforehand, you passed the model in and the initialiser still
queried the db directly.