Always refresh_materialized_views() before building static site
We have recently noticed that the dashboards published on the static
site are slightly behind the same dashboard in the publisher.

We suspect this may be because the materialized views used to build the
dashboards are out of date when the site is built immediately after a
new age has been published.  The views are refreshed nightly at 1am, so
the day following publication the Publisher dashboard includes the new
pages, but the static site still has the old version.

Forcing the views to be refreshed before starting any build of the site
should fix this issue.