Add data correction question
When a minor update is published for a measure page, we want to know
whether or not the update is being published in order to correct a
mistake in the data so that we can be reputable and responsible
publishers of statistical data.

This patch adds a radio field to the measure page form which is only
required and visible for minor updates asking the user to confirm
whether or not they're correcting a mistake in the data. We can then
have a page elsewhere on the site that lists all of the times we've
corrected data mistakes.

 ## Ticket
https://trello.com/c/6DCYrLBY/1299