Use `InputRequired` instead of `DataRequired`
`InputRequired` ensures that the user has submitted _some_ data, but
doesn't care what that data is. `DataRequired` enforces that the data is
also 'truthy'. Therefore, in order to accept falsey values (e.g. a
literal boolean False), we need to use `InputRequired` instead. We want
to ask whether an update to a measure version is due to correcting a
mistake in the data, and this is stored in the database as a boolean, so
we need to be able to accept falsey values in forms.