Fix logic for frequency of release 'other' choice
The data source form used the DB id of the frequency of release records
to look up the index of the 'other' value. This is fine for our
production database where we have a set of specific ordered entries, but
in tests we insert random frequencies of release or records without
reliable IDs. What we actually want is to check whether the user's
selected option is "other" - which we accept as a magic string for the
time being).