Fix FactoryBoy build/create strategies
FactoryBoy provides two strategies for generating data: `build` and
`create`. `build` creates an instance but does persist it to the
database (or register it in the session); `create` adds it to the
session and persists it according to the
`sqlalchemy_session_persistence` setting (which we set to `commit`).

This allows us to construct in-memory instances using the
`SomeFactory.build()` and construct database-committed instances using
`SomeFactory.create()`. Previously, this wasn't configured correctly to
allow the build strategy to flow along relationships, which meant the
create/persist strategy was generally disabled, with everything just
treating it as an in-memory build. This has now been resolved, allowing
us to clearly decide which kind of instance we want.

The default strategy (used when you simply instantiate the class with
`SomeFactory()`) is to commit it to the database. This is the safest
option. If you know that you don't need it persisted, or you're only
generating it to access some faked values and then discard the object,
you can use the `.build()` variant.