Show latest published version to department users (#942)
Currently, department users can only see measure they've been explicitly given access to edit. However it would be useful for them to also be able to see the content of published measure versions when browsing the publisher.

This changes means that when navigating to the `/latest` pages for measures, department users will see the 'latest published version' for that measure, rather than a 403 error or a draft version.

When browsing to specific version pages (eg `/2.1`) the content is visible to the department user if it has been published, but otherwise shows a 403 unauthorised error.