Create command to copy uploads
We want to read/write uploads to keys based on the measure id rather
than the measure/version guid, which is an arbitrary string. This will
allow us to retire guid altogether in the future.

This patch adds a management command that can be run at any time to copy
all keys under guid prefixes to the same bucket, but with id prefixes.