Remove all deprecated data model refs
Add a migration that will finally remove all of the old, now-unused data
model fields after we've done some normalisation work. This migration
*deletes* some data, so reversing it would be pretty difficult. However,
we are currently running the site entirely from the new data model, so
there should be limited risk that we still carry.

A lot of this patch is cleaning up tests so that they no longer
generate or use guids for measure versions.

 ## Ticket
https://trello.com/c/7LiYYoyb/1239