Hide custom classification panel properly
Chartbuilder had a `classList.remove` for both sides of the conditional.
We actually want to add the `hidden` class if the classification is not
custom. Also, both chart and table builders didn't correctly call code
to show/hide the custom classification panel after populating it with
data from the server.

 ## Ticket
https://trello.com/c/wrufHlxo/1328