Fix `get_latest_version_of_all_measures`
The logic for this method was faulty - it would always filter on the
`published` attribute on the model regardless of the value of
`include_drafts`. We only want to filter out drafts if `include_drafts`
is false, otherwise we want to apply no filtering. It's also slightly
wonky to filter on `published`; instead we want to include/exclude by
whether the measure is approved (as it should show up in results if it's
been approved but just hasn't yet made it out to the static site).
Because of this, it feels wrong to call the parameter `include_drafts` -
instead, we rename it to `include_not_published`, which feels like a more
accurate description now.

 ## Ticket
https://trello.com/c/lzMBSpbP/1322