Include testing topic in edit measure page
The testing topics no longer appear in the dropdown on the edit measure
page that allows users to move measures between topics/subtopics. This
was caught yesterday in the department training session, when users
created topics in the testing space. When they came to save them,
because the testing space subtopics weren't there, the dropdown chose an
arbitrary alternative, which moved the measures into the public side of
the site. Testing space topics will now appear in the dropdown once
more.

 ## Ticket
https://trello.com/c/VmfSTXOm/1350