Switch margin to padding
Using padding instead of margin on the bottom of each list item means that the border (which appears when dragging) appears beneath the whitespace rather than before it.

It also means there are no 'gaps' between the list items, which were causing a problem if you dragged-and-dropped an item into the gap (as then the event target is the whole list itself, and so you don't know where to move the item to).