Add env var for staging follower
Review apps need to pull data in from staging. Previously this opaquely
connected to a specificly-named database instance on Heroku's staging
app. We'll clearly name this variable so it can be updated in the
future, though this will require duplicating the credential manually. I
don't see a good alternative for now.