Scroll past sticky banner
When using validation error anchors to jump to the problematic question,
we want to aim to have the label+question fully visible in the viewport.
This effort is frustrated by the sticky banner that we have on the edit
measure page.

This javascript adds a scroll after clicking the validation links which
calculates the height of the sticky header, if present, and scrolls past
it so that the label should be fully visible.

https://trello.com/c/16fidoGF/1359