Make some Chart and Table columns nullable
We want to make sure every Dimension with a chart has a dimension_chart
and every Dimension with a table has a dimension_table.

But we don't know the correct classification for these charts and tables
so these fields will have to be nullable.

This also fixes model specs for `classification_id`, which is currently
defined as an integer but is actually a VARCHAR(255).