Data migration to move charts and tables out of Dimension
Not all charts and tables convert smoothly with our existing v1_to_v2
code, and this script will just skip the failures.

There are currently 3 tables and 2 charts that fail to convert, and
these are all on dimensions that belong to measures that already have
published updates.  So it's fine if we no longer have builder settings
for these, as they will never feature in any future update.

The script is cautious in how it copies objects across - it will not
overwrite any existing values, and only fills in NULL fields.  This
means we should be able to safely run the script more than once if needs
be, without risking overwriting more recent versions of charts or tables.