Use `latest` version reference instead of number
When we publish measure versions, we publish the latest minor update to
each major version under it's numerical identifier (e.g. 1.2), with the
exception of the *very* latest version, which is only available under
`latest`. Therefore, when creating links from the corrections page, we
need to use the `latest` identifier if there are no other versions which
have been published afterwards in order to get a valid URL.

An alternative solution is to duplicate the latest published version and
allow it to be referenced both by its numerical id and the word
`latest`, but this felt quicker to get out and less likely to lead to
dead links hanging around in the future.

https://trello.com/c/3zkveCER/1383