Allow non-uppercase N/A in tables
Sometimes we don't have data for certain cells in a table, and in some
of those cases we use N/A to represent this lack of data. When N/A is
entered in capitals, we add a footnate explaining what this means. But
this check is case-sensitive, so sometimes when 'n/a' is entered, this
footnate isn't generated. This patch will make sure that the footnate is
shown for all case variations of 'N/A'.