Cascade delete of Dimension to its Chart and Table
This also updates other cascades from `all, delete` to
`all, delete-orphan`.

This seems to be what the original intention was; `delete` is included
in `all` anyway.

See the docs here:
https://docs.sqlalchemy.org/en/latest/orm/cascades.html

Which says "[`delete` and `delete-orphan` are appropriate for related
objects which only exist as long as they are attached to their parent,
and are otherwise deleted."