BUGFIX: Chart or table classifications could be NULL
Up until recently if there was a dimension_chart or dimension_table
object associated with a Dimension then they were guaranteed to include
a classification (as the classification fields were not nullable).

Since the addition of more fields to dimension_chart and dimension_table
and the migration of chart and table object data to those tables it is
possiblr for the classification to be null.

This adds a check that the classification exists before trying to
dereference `xxx.classification.ethnicities_count`, which prevents
hitting
 ```
 AttributeError: 'NoneType' object has no attribute 'ethnicities_count'
 ```