Fix file upload and edit flows
Currently if users try to edit the title or description of an upload
they are forced to re-upload a data file.  If a file has already been
uploaded then users shouldn't have to re-upload a file just to tweak
the description.

However, when creating a new upload users should be forced to add a
file.  The NewUploadForm added here enforces this (which is the current
behaviour) but the UploadForm for editing uploads no longer has the
validator on the upload field.

This also adds styling to the error message shown when the user fails
to add a file when creating an upload.