Add data migration to drop old values
Drop chart and table titles from the chartFormat/tableValues blob as
this information is now stored in a dedicated column on each table
(`title`).

This makes use of the #- jsonb functionality in a Postgres database,
which drops the key found at the specified path. In order to use this,
we have to convert our JSON data into Postgres's JSONB format. This does
a few things:

* Removes irrelevant whitespace, eg between keys and values,
* Removes ordering information on keys
* Drops earlier instances of duplicated keys

None of these should cause us any issues, but it feels good to still be
aware that this happens during the process. If we wanted to avoid the
use of jsonb, which would prevent using the #- shorthand, we'd need to
send all the existing chart/table settings through to Python and then
drop the keys there, then sync them back to Postgres.

 ## Ticket
https://trello.com/c/mNdo5JEE