Return "remove secondary source" button
The javascript controlling the 'remove source' button broke due to
trying to add two classes simultaneously via a method that is designed
to add classes one at a time.

After fixing this, the styling on the button was visibly broken, and
`govuk-link` isn't made to override the default button styling, so the
element is now just a link (hooked into a JS function).

A functional test has been added to validate that you can
add/edit/delete secondary sources.

 ## Ticket
https://trello.com/c/9fo6P53K