Mgmt command to drop measures from database
Heroku only supports up to 10k rows on review app databases at the
moment, which is a threshold we've now exceeded. In order to remain
below the limit, we're going to try dropping all measures except the
first two in each subtopic. This should give us a decent enough sample
to aid in reviewing, without taking us over the 10k row limit.