Include preferred capitalisation for all DB Ethnicities
We've seen a recent problem where "Irish traveller" in source data is
not mapped to "Irish Traveller" as an ethnicity and so finding
classifications for a chart/table fails.

This ensures that there is an entry in `classification_lookup.csv` for
every ethnicity that we have in our database. (It turns out there were
a few missing.)