Convert details component to use a macro (#993)
This adds a Jinja2 macro for the Details component from the GOV.UK Design System, and converts our existing templates to use the new macro.

The main advantages are in not having to remember and re-type all the nested class names and elements.

The macro itself is based on the Nunjucks one from govuk-frontend, but converts the syntax to Jinja2. All the param names are the same.