Fix: only add buttons to fixed table header
This fixes a bug whereby if there is a fixed table header (which duplicates the table `thead` into a new table so that the fixed table can be positioned over the overflow-scrolling table), the sort buttons are ONLY added to the fixed header.

This mainly caused an issue to keyboard users, as the invisible duplicate sortable table buttons could still be tabbed to.