Add a test to detect db_version_id increment
We have a column on our MeasureVersion model that is meant to track the
number of updates applied to each row. This helps us identify when two
users are editing a measure at the same time. If both users load the
same version of the measure, and then they both save some changes
sequentially, the later update risks overwriting the earlier update. By
tracking each saved version, we can detect when a save might overwrite
changes, and play the changes back to the end user, giving them an
opportunity to rationalise their changeset with the other person's.

Unfortunately db_version_id has not been incrementing correctly, and we
had no test coverage to warn us. This test should provide confidence in
the future that this functionality works.

 ## Ticket
https://trello.com/c/eXCUGoTs