Fix db_version_id tracking
A combination of factors led to the loss of an auto-incremented
`db_version_id`, mostly centred around the move to manually populating
each field on the model from the form to use
`form.populate_obj(model_instance)`.

The WTForm field for `db_version_id` was a `HiddenField`, which uses
string types. `db_version_id` on the database is an integer field. When
populating a numeric string into the SQLAlchemy model instance, upon
save, the version is not incremented (but the numeric-string is
automatically cast to an int, so no error is thrown). By switching to an
`IntegerField` in WTForms (using a hidden widget, so that it isn't
visible on the page)`, we get the value cast to a Python int
automatically. When this is injected into the model instance, and
flushed to the database, it *does* get incremented properly.

With just that update in place, `db_version_id` started get incremented
again. But on some of the tests, it would be incremented *twice*. It
turns out that this is because we were manually assigning the title to
the measure version, then running some queries to check whether the slug
needed updating, which caused an implicit flush to the database, saving
the new title and incrementing the ID. After this check, we then
populated the rest of the fields to the measure version and updated it
again, which bumped `db_version_id` for a second time. Given that the
title is already injected into the measure version by the second stage,
we can safely remove the initial manual set.

 ## Ticket
https://trello.com/c/eXCUGoTs