Test conflicting updates
We were missing a test that validates the user experience when two users
are updating the same measure around the same time. If user one and two
are both working on the same measure, and user one saves it, when user
two comes to save it they should be presented with a validation screen
that displays the diff between user one's changes and their own. They
can then manually perform a merge to bring the measure back to an ideal
state.

When the user is presented with a set of diffs, the new `db_version_id`
needs to be injected into the page so that a subsequent save will not
keep presenting the same errors.

 ## Ticket
https://trello.com/c/eXCUGoTs