Don't fake `db_version_id`
This column is hooked into SQLAlchemy, and we rely on it for tracking
changes to our model. If we fake it, we risk not knowing when this
breaks. It also isn't easy to fake, as SQLAlchemy can overwrite what you
specify through FactoryBoy with its own version of the truth (e.g. upon
initial creation, where it's always set to 1).

 ## Ticket
https://trello.com/c/eXCUGoTs