Consistent h3 and link headings
On the homepage we have two central sections of the page: links to
ethnicity data by topic, and links to our updates dashboards.

The ethnicity data by topic links are not headings, but the dashboard
links are `h3`s. We should probably be consistent so that screen-reading
users have an easier time navigating around the site. GOV.UK's homepage
has their equivalent of what we have as `h3`s in an unordered list, so
it feels safe to take advantage of the work they've done and copy their
pattern.