Update other password forms to RDU fields
While doing the 'Forgot password' page I also noticed that the other
password-related forms were not using the RDU fields, so I've converted
those as well. The `SetPasswordForm` form covers both resetting a
password if the user has forgotten it, as well as setting the password
upon initial account creation.