Only show Update button on latest versions
We recently made a change to allow any version of a measure to be
unpublished:
https://github.com/racedisparityaudit/ethnicity-facts-and-figures-publisher/commit/b2bddac54c47d5f7b2d4b2bd2d48f334981e7355

But at the same time we inadvertently added an "Update" button to all
versions of a measure, but this is a thing that should only appear on
the latest version.

The backend already handles the case where a request is made to update
an old version (with the message "Version x.y of page zzzz is already
being updated").

But showing the button is confusing and we should hide it for cases when
it won't work.