Mark pages unpublished before build
At the moment it takes two static builds for a page to be removed from
the live static site: the first build marks the page as unpublished, but
still builds and pushes the page up, and then a second build recognises
the page has been unpublished and actually removes it from the live
site. We should mark pages unpublished and remove them at the same time
in the first static build, which is what this patch aims to do.

 ## Ticket
https://trello.com/c/hHtanMW1