Also log raw request data for updates
We have recently seen another case where measure data was inexplicably
lost.

The logging we have allowed us to see that the WTForm data used to
update the measure was mostly empty, but we don't know if the data was
lost in the browser or inside our app somehow.

By also logging the request data that came in to Flask if this ever
happens again we might be able to figure out a bit more.