Fix classification options in custom
Fixes a bug that led to empty dropdowns where a full list of our known
classifications should be provided to the user. The classifications were
not being passed into the template and no tests were in place to detect
this - we now have one.

I've created a tech improvement ticket from this
(https://trello.com/c/fwYpIWkD) that records the fact our tests are
currently using 'real' data rather than the test data for classification
finders/classications. Our tests break if we switch the config for this,
so leaving it to a later ticket to deal with separately.

 ## Ticket
https://trello.com/c/gRGCp1yr