Enable strict undefined template processing
At the moment, Jinja2 handles undefined variables as if they were
`None`s. This can make it nice and quick to write templates, as you
don't need good error-checking or handling, but we've introduced a fair
few bugs lately that would have been caught by the rigour imposed by
undefined variables bubbling up as errors. In the spirit of failing
fast, and preventing future bugs, we should turn this on.