Pass `static_mode` template parameter consistently
At the moment, we accept the flag `static_mode` into the templating
engine in a fairly sporadic way - specifically in each `render_template`
call when we know we need to do it in one of two ways. We also accept
the `static_mode` URL query param on some pages, but not others. By
adding it into Jinja's context processor, we can cleanly and easily work
out if the page should be rendered as static by checking the current
app's config, and then also checking the current request's parameters.
It also means that the variable is defined for every template render
call, so we don't need to test whether it's defined or not before
accessing.