Remove chart, add tables to dimensions
With `StrictUndefined`, we need slightly more realistic test data when
testing out the static site build. Because we have some conditions that
switch on the presence of a 'dimension_chart' or 'dimension_table'
relationship, it's expected that if a record exists it can be rendered.
For this to be true, it needs a `chart/table_object` JSON blob. Let's
make sure there's no chart for the given dimensions, but provide at
least a table - which for us, is expected to always be provided.