Enhance JS fixed tables with captions
Adds behaviour to the fixed table header functionality that allows it to
also display the table title within the fixed area. It does this by
duplicating the table caption into the fixed header `table` element and
setting the original caption to be transparent. The fixed caption then
sits over the top of the original one.

 ## Ticket
https://trello.com/c/gxKxOj16