Update node library versions and dependencies
GitHub is reporting a high severity vulnerability in gulp-sass which may
(or may not) be fixed by updating all the things here. I suspect maybe
not because `npm audit` is still reporting the vulnerability.

If it isn't fixed here then I think we have to wait for packages
upstream to patch their dependencies before we can.

But it probably doesn't hurt to update this lot while we're here anyway.