Bugfix: set publication dates on published measures
This fixes a bug in a previous script to update measures with new dimension titles (from a CSV file).

That script assumed that the build process would set a publication date, whereas this actually seems to require `mark_measure_version_published` to be called (which is normally done at time of pressing 'Approve').